feature/saving-recipe in release-1.10

+++ GroceryList/ViewControllers/MainViewController/Childs/RecipesListViewController/Childs/RecipeViewController/RecipeViewController.swift
@@ -307,7 +307,7 @@ final class RecipeViewController: UIViewController {
 }
 
 extension RecipeViewController: RecipeServingSelectorDelegate {
-    func servingChangedTo(count: Int) {
+    func servingChangedTo(count: Double) {
         let servings = viewModel.getIngredientsSizeAccordingToServings(servings: count)
         updateIngredientsCount(by: servings)
     }
